Modularise lib
