+++ src/redqct/image.py
@@ -6,7 +6,8 @@ from fontTools.ttLib import TTFont
 
 from discord import Status, Colour
 
-from .lib import MemberAttrs, fetch_all
+from .lib import MemberAttrs
+from .lib.http import fetch_all
 
 
 ROOT_DIR = Path(__file__).resolve().parents[2]
